remove mandatory creation of config varibales
since I try to use null paterns for configuration variables, registering it does not make any sense now

+++ EmbUI/basicui.h
@@ -21,7 +21,7 @@ extern uint8_t lang;
 namespace basicui {
 
     // numeric indexes for pages
-    enum class page : uint8_t {
+    enum class page : uint16_t {
         main = 0,
         settings,
         network,
